# openapi.yaml
openapi: 3.0.3
info:
  title: Lottery API
  version: 1.0.0
  description: Small API for adding and listing lottery entries.

paths:
  /entries:
    get:
      summary: List lottery entries
      operationId: listEntries
      tags:
        - entries
      responses:
        "200":
          description: List of all lottery entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LotteryEntry"
    post:
      summary: Add lottery entry
      operationId: addEntry
      tags:
        - entries
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddEntryRequest"
      responses:
        "201":
          description: Lottery entry created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LotteryEntry"
        "400":
          description: Invalid request (missing name)
        "409":
          description: Name already registered
  /draw:
    get:
      summary: Draw lottery
      operationId: drawLottery
      tags:
        - draw
      responses:
        "200":
          description: Winner selected and entries reset
          content:
            application/json:
              schema:
                type: object
                properties:
                  winner:
                    $ref: "#/components/schemas/LotteryEntry"
                  message:
                    type: string
        "404":
          description: No entries to draw

components:
  schemas:
    LotteryEntry:
      type: object
      properties:
        id:
          type: string
          description: Unique entry identifier
        name:
          type: string
          description: Name of the participant
        ticket:
          type: string
          description: Lottery ticket number or code
        created_at:
          type: string
          format: date-time
          description: When the entry was created
    AddEntryRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          description: Unique participant name; ticket is generated by the API